+++ src/pipelayer/__init__.py
@@ -4,5 +4,4 @@ from pipelayer.event_args import FilterEventArgs  # NOQA F401
 from pipelayer.filter import Filter  # NOQA F401
 from pipelayer.manifest import Manifest  # NOQA F401
 from pipelayer.pipeline import Pipeline  # NOQA F401
-from pipelayer.step import StepHelper  # NOQA F401
 from pipelayer.switch import Switch  # NOQA F401