+++ test/unit/test_step_helper_functions.py
@@ -1,31 +1,32 @@
 import pytest
-from pipelayer import StepHelper
 from pipelayer.exception import InvalidFilterException
+from pipelayer.step import (initialize_step, is_callable_valid,
+                            is_class_func_static)
 
 
 @pytest.mark.unit
-class TestStepHelper:
+class TestStepHelperFunctions:
 
     @pytest.mark.sad
     def test_initialize_step_none_step(self):
         with pytest.raises(InvalidFilterException):
-            StepHelper.initialize_step(None)
+            initialize_step(None)
 
     @pytest.mark.sad
     def test_initialize_step_invalid_step(self):
         with pytest.raises(InvalidFilterException):
-            StepHelper.initialize_step(lambda a, b, c: a + b + c)
+            initialize_step(lambda a, b, c: a + b + c)
 
     @pytest.mark.sad
     def test_is_callable_valid(self):
         def func(data, Context):
             return True
-        assert StepHelper.is_callable_valid(func)
+        assert is_callable_valid(func)
 
     @pytest.mark.sad
     def test_is_callable_valid_false(self):
-        assert StepHelper.is_callable_valid(None) is False
-        assert StepHelper.is_callable_valid("test") is False
+        assert is_callable_valid(None) is False
+        assert is_callable_valid("test") is False
 
     @pytest.mark.sad
     def test_is_class_func_static_false(self):
@@ -33,7 +34,7 @@ class TestStepHelper:
             def run(self, data, context):
                 pass
 
-        assert StepHelper._StepHelper__is_class_func_static(MyClass) is False
+        assert is_class_func_static(MyClass) is False
 
     @pytest.mark.sad
     def test_is_class_func_static_type_error(self):
@@ -41,7 +42,7 @@ class TestStepHelper:
             pass
 
         with pytest.raises(AttributeError):
-            StepHelper._StepHelper__is_class_func_static(MyClass)
+            is_class_func_static(MyClass)
 
     @pytest.mark.happy
     def test_is_class_func_static(self):
@@ -52,7 +53,7 @@ class TestStepHelper:
             def run(data, context):
                 pass
 
-        assert StepHelper._StepHelper__is_class_func_static(MyClass)
+        assert is_class_func_static(MyClass)
 
     @pytest.mark.happy
     def test_is_class_func_static_using_instance(self):
@@ -63,4 +64,4 @@ class TestStepHelper:
             def run(data, context):
                 pass
 
-        assert StepHelper._StepHelper__is_class_func_static(type(MyClass()))
+        assert is_class_func_static(type(MyClass()))